fetch the current time in UTC
The WiFi example code in the library included an example of fetching
the current time from time.nist.gov. I modified this to fetch from
time.google.com instead, and I moved the network credentials to a
separate file to avoid committing them to the repository.

This example seems to take up about 70% of the available storage on
my device; I hope that the other libraries I want to use don't make
it take up more than I have available! Or that I can use the SD
card slot to store a larger program on my device.