WiFi IPV6 code which fetches the time
This code works as is, but includes some very long delays, which
means that all of the other inputs on the rotary switch and the
twist have a ~10s wait to see the result. Obviously, this won't
do.

The next step is to redesign the WiFi loop function to always
end quickly and check the elapsed time instead of forcing the
main thread to idle for long periods.