Track 2 previous mode values and use them to ignore analog noise.
While testing "night" mode, I noticed that occasionally the lights
would flicker briefly because they were flashing for one moment into
a mode with nearby adjacent analog values for the switch. That is,
every once in a while the analog read values for the switch would be
slightly out of range. In order to resolve this, I keep a small
buffer of the previous two readings. This allows me to ignore an
isolated value which is out of range and doesn't reflect an actual
switch.