Implement a "sliding window" brightness control
The Twist Rotational input can boot up with "any" value to start.
As you spin the dial, the original algorithm was computing brightness
and then limiting it to 0-255. But you could keep going outside of
the range, and you couldn't go the opposite direction until you
went all the way back to within the range.

This is solved by making a sliding window: if you are already at
the maximum brightness, turning the dial clockwise will slide the
window up. So when you turn the dial counterclockwise, the brightness
will immediately start to reduce. This feels more intuitive.